refactor: Refactor RecipeListPage
