+++ src/components/modules/recipes-list/RecipeListBanner/assets/bannerrecipes.png
