+++ src/components/modules/recipes-list/index.ts
@@ -1,0 +1,2 @@
+export { default as RecipeListBanner } from './RecipeListBanner/RecipeListBanner'
+export { default as RecipesList} from './RecipesList/RecipesList'
