--- src/components/modules/recipes/RecipesList/RecipesList.module.scss
@@ -1,43 +1,0 @@
-@import "../../../../styles/_utilities";
-
-.recipesListWrapper {
-    @screen md{
-        padding-bottom:5.6rem;
-    }
-    .recipesHead{
-        @apply flex justify-end;
-        @screen md{
-            @apply  justify-between;
-        }
-    }
-    .recipesListBlogs{
-        @screen md {
-            @apply flex flex-col items-center justify-center;
-        }
-        .recipesList {
-            @screen md {
-                @apply flex justify-between flex-wrap;
-                margin: 1rem 0;
-            }
-            .recipesItem {
-                @screen md {
-                    width: calc(97% / 2);
-                    margin-top:4rem;
-                }
-                @screen lg{
-                    width: calc(97% / 3);
-                    margin-top:4rem;
-                }
-            }
-        }
-    }
-    .recipesPagination{
-        @apply flex justify-center;
-        margin: 3rem 0;
-        @screen md {
-            @apply flex justify-center;
-            margin:2rem 0;
-        }
-    }
-   
-}
