--- src/components/modules/recipes/index.ts
@@ -1,1 +1,0 @@
-export { default as RecipeListPage } from './RecipeListPage/RecipeListPage'
